Revert "[consolidation] Make variant_task a utask (#3085)"
This reverts commit 2b445033313a292231b37e30fe85405041113c3e.
