+++ src/clusterfuzz/_internal/bot/tasks/utasks/uworker_errors.py
@@ -20,5 +20,5 @@ class Type(enum.Enum):
   ANALYZE_BUILD_SETUP = 1
   ANALYZE_NO_CRASH = 2
   TESTCASE_SETUP = 3
-  UNHANDLED = 4
-  VARIANT_BUILD_SETUP = 5
+  NO_FUZZER = 4
+  UNHANDLED = 5
