+++ src/clusterfuzz/_internal/bot/tasks/utasks/uworker_handle_errors.py
@@ -15,7 +15,6 @@
 from clusterfuzz._internal.bot.tasks import setup
 from clusterfuzz._internal.bot.tasks.utasks import analyze_task
 from clusterfuzz._internal.bot.tasks.utasks import uworker_errors
-from clusterfuzz._internal.bot.tasks.utasks import variant_task
 
 
 def noop(*args, **kwargs):
@@ -25,19 +24,18 @@ def noop(*args, **kwargs):
 
 def handle(output):
   """Handles the errors bubbled up from the uworker."""
-  return get_mapping()[output.error](output)
+  return MAPPING[output.error](output)
 
 
-def get_mapping():
-  return {
-      uworker_errors.Type.ANALYZE_NO_CRASH:
-          analyze_task.handle_noncrash,
-      uworker_errors.Type.ANALYZE_BUILD_SETUP:
-          analyze_task.handle_build_setup_error,
-      uworker_errors.Type.TESTCASE_SETUP:
-          setup.handle_setup_testcase_error,
-      uworker_errors.Type.VARIANT_BUILD_SETUP:
-          variant_task.handle_build_setup_error,
-      uworker_errors.Type.UNHANDLED:
-          noop,
-  }
+MAPPING = {
+    uworker_errors.Type.ANALYZE_NO_CRASH:
+        analyze_task.handle_noncrash,
+    uworker_errors.Type.ANALYZE_BUILD_SETUP:
+        analyze_task.handle_build_setup_error,
+    uworker_errors.Type.TESTCASE_SETUP:
+        setup.handle_setup_testcase_error,
+    uworker_errors.Type.NO_FUZZER:
+        noop,
+    uworker_errors.Type.UNHANDLED:
+        noop,
+}
